extract wave simulation route
